+++ src/deploy-notebook/index.ts
@@ -3,7 +3,7 @@ import { ToolbarButton } from '@jupyterlab/apputils';
 import { DocumentRegistry } from '@jupyterlab/docregistry';
 import { IDisposable, DisposableDelegate } from '@lumino/disposable';
 
-import { showDeploymentDialog } from '../dialog';
+import { showUploadDialog } from '../dialog';
 import { settingsChanged, JupySQLSettings } from '../settings';
 
 /**
@@ -42,19 +42,19 @@ export class DeployingExtension
     ): IDisposable {
 
         const clickDeploy = () => {
-            showDeploymentDialog(panel, context)
+            showUploadDialog(panel, context)
         }
 
         this.panel = panel;
 
         this.deployNotebookButton = new ToolbarButton({
-            className: 'deploy-nb-button',
-            label: 'Deploy Notebook',
+            className: 'share-nb-button',
+            label: 'Share Notebook',
             onClick: clickDeploy,
-            tooltip: 'Deploy Notebook as dashboards',
+            tooltip: 'Share notebook by uploading it to Ploomber Cloud',
         });
 
-        this.deployNotebookButton.node.setAttribute("data-testid", "deploy-btn");
+        this.deployNotebookButton.node.setAttribute("data-testid", "share-btn");
 
         panel.toolbar.insertItem(10, 'deployNB', this.deployNotebookButton);